Upgrading to 1.10.8

+++ tests/Zend/Service/WindowsAzure/QueueStorageTest.php
@@ -15,7 +15,7 @@
  * @category   Zend
  * @package    Zend_Service_WindowsAzure
  * @subpackage UnitTests
- * @version    $Id: QueueStorageTest.php 24241 2009-07-22 09:43:13Z unknown $
+ * @version    $Id$
  * @copyright  Copyright (c) 2005-2010 Zend Technologies USA Inc. (http://www.zend.com)
  * @license    http://framework.zend.com/license/new-bsd     New BSD License
  */
@@ -25,49 +25,42 @@
  */
 require_once dirname(__FILE__) . '/../../../TestHelper.php';
 
-/**
- * @see Zend_Service_WindowsAzure_Storage_Queue 
- */
+/** Zend_Service_WindowsAzure_Storage_Queue */
 require_once 'Zend/Service/WindowsAzure/Storage/Queue.php';
-
-if (!defined('PHPUnit_MAIN_METHOD')) {
-    define('PHPUnit_MAIN_METHOD', 'Zend_Service_WindowsAzure_QueueStorageTest::main');
-}
 
 /**
  * @category   Zend
  * @package    Zend_Service_WindowsAzure
  * @subpackage UnitTests
- * @version    $Id: QueueStorageTest.php 24241 2009-07-22 09:43:13Z unknown $
+ * @version    $Id$
  * @copyright  Copyright (c) 2005-2010 Zend Technologies USA Inc. (http://www.zend.com)
  * @license    http://framework.zend.com/license/new-bsd     New BSD License
  */
 class Zend_Service_WindowsAzure_QueueStorageTest extends PHPUnit_Framework_TestCase
 {
-    public static function main()
-    {
-        $suite  = new PHPUnit_Framework_TestSuite("Zend_Service_WindowsAzure_QueueStorageTest");
-        $result = PHPUnit_TextUI_TestRunner::run($suite);
-    }
-
+    
+    protected static $uniqId = 0;
+       
     /**
      * Test setup
      */
     protected function setUp()
     {
         if (!TESTS_ZEND_SERVICE_WINDOWSAZURE_QUEUE_RUNTESTS) {
-            $this->markTestSkipped('Windows Azure Tests disabled');
+            $this->markTestSkipped('This test case requires TESTS_ZEND_SERVICE_WINDOWSAZURE_QUEUE_RUNTESTS to be enabled in TestConfiguration.php');
         }
     }
-
+    
     /**
      * Test teardown
      */
     protected function tearDown()
     {
+        if ($this->status == PHPUnit_Runner_BaseTestRunner::STATUS_SKIPPED) {
+            return;
+        }
         $storageClient = $this->createStorageInstance();
-        for ($i = 1; $i <= self::$uniqId; $i++)
-        {
+        for ($i = 1; $i <= self::$uniqId; $i++) {
             try { $storageClient->deleteQueue(TESTS_ZEND_SERVICE_WINDOWSAZURE_QUEUE_PREFIX . $i); } catch (Exception $e) { }
         }
     }
@@ -80,22 +73,20 @@ class Zend_Service_WindowsAzure_QueueStorageTest extends PHPUnit_Framework_TestC
         } else {
             $storageClient = new Zend_Service_WindowsAzure_Storage_Queue(TESTS_ZEND_SERVICE_WINDOWSAZURE_QUEUE_HOST_DEV, TESTS_ZEND_SERVICE_WINDOWSAZURE_STORAGE_ACCOUNT_DEV, TESTS_ZEND_SERVICE_WINDOWSAZURE_STORAGE_KEY_DEV, true, Zend_Service_WindowsAzure_RetryPolicy_RetryPolicyAbstract::retryN(10, 250));
         }
-
+        
         if (TESTS_ZEND_SERVICE_WINDOWSAZURE_STORAGE_USEPROXY) {
             $storageClient->setProxy(TESTS_ZEND_SERVICE_WINDOWSAZURE_STORAGE_USEPROXY, TESTS_ZEND_SERVICE_WINDOWSAZURE_STORAGE_PROXY, TESTS_ZEND_SERVICE_WINDOWSAZURE_STORAGE_PROXY_PORT, TESTS_ZEND_SERVICE_WINDOWSAZURE_STORAGE_PROXY_CREDENTIALS);
         }
 
         return $storageClient;
     }
-
-    protected static $uniqId = 0;
-
+    
     protected function generateName()
     {
         self::$uniqId++;
         return TESTS_ZEND_SERVICE_WINDOWSAZURE_QUEUE_PREFIX . self::$uniqId;
     }
-
+    
     /**
      * Test queue exists
      */
@@ -109,11 +100,11 @@ class Zend_Service_WindowsAzure_QueueStorageTest extends PHPUnit_Framework_TestC
 
         $result = $storageClient->queueExists($queueName1);
         $this->assertTrue($result);
-
+        
         $result = $storageClient->queueExists(md5(time()));
         $this->assertFalse($result);
     }
-
+    
     /**
      * Test create queue
      */
@@ -124,7 +115,7 @@ class Zend_Service_WindowsAzure_QueueStorageTest extends PHPUnit_Framework_TestC
         $result = $storageClient->createQueue($queueName);
         $this->assertEquals($queueName, $result->Name);
     }
-
+    
     /**
      * Test set queue metadata
      */
@@ -133,15 +124,15 @@ class Zend_Service_WindowsAzure_QueueStorageTest extends PHPUnit_Framework_TestC
         $queueName = $this->generateName();
         $storageClient = $this->createStorageInstance();
         $storageClient->createQueue($queueName);
-
+        
         $storageClient->setQueueMetadata($queueName, array(
             'createdby' => 'PHPAzure',
         ));
-
+        
         $metadata = $storageClient->getQueueMetadata($queueName);
         $this->assertEquals('PHPAzure', $metadata['createdby']);
     }
-
+    
     /**
      * Test get queue
      */
@@ -150,12 +141,12 @@ class Zend_Service_WindowsAzure_QueueStorageTest extends PHPUnit_Framework_TestC
         $queueName = $this->generateName();
         $storageClient = $this->createStorageInstance();
         $storageClient->createQueue($queueName);
-
+        
         $queue = $storageClient->getQueue($queueName);
         $this->assertEquals($queueName, $queue->Name);
         $this->assertEquals(0, $queue->ApproximateMessageCount);
     }
-
+    
     /**
      * Test list queues
      */
@@ -175,13 +166,31 @@ class Zend_Service_WindowsAzure_QueueStorageTest extends PHPUnit_Framework_TestC
         $storageClient->deleteQueue($queueName1);
         $storageClient->deleteQueue($queueName2);
         $storageClient->deleteQueue($queueName3);
-
+        
         $this->assertEquals(3, count($result1));
         $this->assertEquals($queueName2, $result1[1]->Name);
-
+        
         $this->assertEquals(1, count($result2));
     }
-
+    
+    /**
+     * Test list queues with metadata
+     */
+    public function testListQueuesWithMetadata()
+    {
+        $queueName = $this->generateName();
+        $storageClient = $this->createStorageInstance();
+        $storageClient->createQueue($queueName, array(
+            'createdby' => 'PHPAzure',
+            'ownedby' => 'PHPAzure',
+        ));
+        
+        $result = $storageClient->listQueues($queueName, null, null, 'metadata');
+        
+        $this->assertEquals('PHPAzure', $result[0]->Metadata['createdby']);
+        $this->assertEquals('PHPAzure', $result[0]->Metadata['ownedby']);
+    }
+    
     /**
      * Test put message
      */
@@ -191,14 +200,14 @@ class Zend_Service_WindowsAzure_QueueStorageTest extends PHPUnit_Framework_TestC
         $storageClient = $this->createStorageInstance();
         $storageClient->createQueue($queueName);
         $storageClient->putMessage($queueName, 'Test message', 120);
-
-        sleep(45); // wait for the message to appear in the queue...
-
+        
+        sleep(5); // wait for the message to appear in the queue...
+        
         $messages = $storageClient->getMessages($queueName);
         $this->assertEquals(1, count($messages));
         $this->assertEquals('Test message', $messages[0]->MessageText);
     }
-
+    
     /**
      * Test get messages
      */
@@ -211,18 +220,18 @@ class Zend_Service_WindowsAzure_QueueStorageTest extends PHPUnit_Framework_TestC
         $storageClient->putMessage($queueName, 'Test message 2', 120);
         $storageClient->putMessage($queueName, 'Test message 3', 120);
         $storageClient->putMessage($queueName, 'Test message 4', 120);
-
-        sleep(45); // wait for the messages to appear in the queue...
-
+        
+        sleep(5); // wait for the messages to appear in the queue...
+        
         $messages1 = $storageClient->getMessages($queueName, 2);
         $messages2 = $storageClient->getMessages($queueName, 2);
         $messages3 = $storageClient->getMessages($queueName);
-
+        
         $this->assertEquals(2, count($messages1));
         $this->assertEquals(2, count($messages2));
         $this->assertEquals(0, count($messages3));
     }
-
+    
     /**
      * Test peek messages
      */
@@ -235,16 +244,39 @@ class Zend_Service_WindowsAzure_QueueStorageTest extends PHPUnit_Framework_TestC
         $storageClient->putMessage($queueName, 'Test message 2', 120);
         $storageClient->putMessage($queueName, 'Test message 3', 120);
         $storageClient->putMessage($queueName, 'Test message 4', 120);
-
-        sleep(45); // wait for the messages to appear in the queue...
-
+        
+        sleep(5); // wait for the messages to appear in the queue...
+        
         $messages1 = $storageClient->peekMessages($queueName, 4);
         $messages2 = $storageClient->getMessages($queueName, 4);
-
+        
         $this->assertEquals(4, count($messages1));
         $this->assertEquals(4, count($messages2));
     }
-
+    
+    /**
+     * Test dequeuecount
+     */
+    public function testDequeueCount()
+    {
+        $queueName = $this->generateName();
+        $storageClient = $this->createStorageInstance();
+        $storageClient->createQueue($queueName);
+        $storageClient->putMessage($queueName, 'Test message 1', 120);
+        
+        sleep(5); // wait for the message to appear in the queue...
+        
+        $expectedDequeueCount = 3;
+        for ($i = 0; $i < $expectedDequeueCount - 1; $i++) {
+            $storageClient->getMessages($queueName, 1, 1);
+            sleep(3);
+        }
+        
+        $messages = $storageClient->getMessages($queueName, 1);
+        
+        $this->assertEquals($expectedDequeueCount, $messages[0]->DequeueCount);
+    }
+    
     /**
      * Test clear messages
      */
@@ -257,20 +289,20 @@ class Zend_Service_WindowsAzure_QueueStorageTest extends PHPUnit_Framework_TestC
         $storageClient->putMessage($queueName, 'Test message 2', 120);
         $storageClient->putMessage($queueName, 'Test message 3', 120);
         $storageClient->putMessage($queueName, 'Test message 4', 120);
-
-        sleep(45); // wait for the messages to appear in the queue...
-
+        
+        sleep(5); // wait for the messages to appear in the queue...
+        
         $messages1 = $storageClient->peekMessages($queueName, 4);
         $storageClient->clearMessages($queueName);
-
-        sleep(45); // wait for the GC...
-
+        
+        sleep(5); // wait for the GC...
+        
         $messages2 = $storageClient->peekMessages($queueName, 4);
-
+        
         $this->assertEquals(4, count($messages1));
         $this->assertEquals(0, count($messages2));
     }
-
+    
     /**
      * Test delete message
      */
@@ -283,24 +315,20 @@ class Zend_Service_WindowsAzure_QueueStorageTest extends PHPUnit_Framework_TestC
         $storageClient->putMessage($queueName, 'Test message 2', 120);
         $storageClient->putMessage($queueName, 'Test message 3', 120);
         $storageClient->putMessage($queueName, 'Test message 4', 120);
-
-        sleep(45); // wait for the messages to appear in the queue...
-
+        
+        sleep(5); // wait for the messages to appear in the queue...
+        
         $messages1 = $storageClient->getMessages($queueName, 2, 10);
-        foreach ($messages1 as $message) {
+        foreach ($messages1 as $message)
+        {
             $storageClient->deleteMessage($queueName, $message);
         }
-
-        sleep(45); // wait for the GC...
-
+        
+        sleep(5); // wait for the GC...
+        
         $messages2 = $storageClient->getMessages($queueName, 4);
-
+        
         $this->assertEquals(2, count($messages1));
         $this->assertEquals(2, count($messages2));
     }
-}
-
-// Call Zend_Service_WindowsAzure_QueueStorageTest::main() if this source file is executed directly.
-if (PHPUnit_MAIN_METHOD == "Zend_Service_WindowsAzure_QueueStorageTest::main") {
-    Zend_Service_WindowsAzure_QueueStorageTest::main();
 }
