+++ tests/Zend/Service/WindowsAzure/SessionHandlerTest.php
@@ -15,7 +15,7 @@
  * @category   Zend
  * @package    Zend_Service_WindowsAzure
  * @subpackage UnitTests
- * @version    $Id: BlobStorageTest.php 14561 2009-05-07 08:05:12Z unknown $
+ * @version    $Id$
  * @copyright  Copyright (c) 2005-2010 Zend Technologies USA Inc. (http://www.zend.com)
  * @license    http://framework.zend.com/license/new-bsd     New BSD License
  */
@@ -25,38 +25,27 @@
  */
 require_once dirname(__FILE__) . '/../../../TestHelper.php';
 
-/**
- * @see Zend_Service_WindowsAzure_SessionHandler 
- */
+/** Zend_Service_WindowsAzure_SessionHandler */
 require_once 'Zend/Service/WindowsAzure/SessionHandler.php';
 
-/**
- * @see Zend_Service_WindowsAzure_Storage_Table 
- */
+/** Zend_Service_WindowsAzure_Storage_Table */
 require_once 'Zend/Service/WindowsAzure/Storage/Table.php';
-
-if (!defined('PHPUnit_MAIN_METHOD')) {
-    define('PHPUnit_MAIN_METHOD', 'Zend_Service_WindowsAzure_SessionHandlerTest::main');
-}
 
 /**
  * @category   Zend
  * @package    Zend_Service_WindowsAzure
  * @subpackage UnitTests
- * @version    $Id: BlobStorageTest.php 14561 2009-05-07 08:05:12Z unknown $
+ * @version    $Id$
  * @copyright  Copyright (c) 2005-2010 Zend Technologies USA Inc. (http://www.zend.com)
  * @license    http://framework.zend.com/license/new-bsd     New BSD License
  */
 class Zend_Service_WindowsAzure_SessionHandlerTest extends PHPUnit_Framework_TestCase
 {
+    
+    protected static $uniqId = 0;
+    
     public function __construct()
     {
-    }
-
-    public static function main()
-    {
-        $suite  = new PHPUnit_Framework_TestSuite("Zend_Service_WindowsAzure_SessionHandlerTest");
-        $result = PHPUnit_TextUI_TestRunner::run($suite);
     }
 
     /**
@@ -65,22 +54,25 @@ class Zend_Service_WindowsAzure_SessionHandlerTest extends PHPUnit_Framework_Tes
     protected function setUp()
     {
         if (!TESTS_ZEND_SERVICE_WINDOWSAZURE_SESSIONHANDLER_RUNTESTS) {
-            $this->markTestSkipped('Windows Azure Tests disabled');
+            $this->markTestSkipped('This test case requires TESTS_ZEND_SERVICE_WINDOWSAZURE_SESSIONHANDLER_RUNTESTS to be enabled in TestConfiguration.php');
         }
     }
-
+    
     /**
      * Test teardown
      */
     protected function tearDown()
     {
+        if ($this->status == PHPUnit_Runner_BaseTestRunner::STATUS_SKIPPED) {
+            return;
+        }
         $storageClient = $this->createStorageInstance();
         for ($i = 1; $i <= self::$uniqId; $i++)
         {
             try { $storageClient->deleteTable(TESTS_ZEND_SERVICE_WINDOWSAZURE_SESSIONHANDLER_TABLENAME_PREFIX . $i); } catch (Exception $e) { }
         }
     }
-
+    
     protected function createStorageInstance()
     {
         $storageClient = null;
@@ -89,14 +81,14 @@ class Zend_Service_WindowsAzure_SessionHandlerTest extends PHPUnit_Framework_Tes
         } else {
             $storageClient = new Zend_Service_WindowsAzure_Storage_Table(TESTS_ZEND_SERVICE_WINDOWSAZURE_TABLE_HOST_DEV, TESTS_ZEND_SERVICE_WINDOWSAZURE_STORAGE_ACCOUNT_DEV, TESTS_ZEND_SERVICE_WINDOWSAZURE_STORAGE_KEY_DEV, true, Zend_Service_WindowsAzure_RetryPolicy_RetryPolicyAbstract::retryN(10, 250));
         }
-
+        
         if (TESTS_ZEND_SERVICE_WINDOWSAZURE_STORAGE_USEPROXY) {
             $storageClient->setProxy(TESTS_ZEND_SERVICE_WINDOWSAZURE_STORAGE_USEPROXY, TESTS_ZEND_SERVICE_WINDOWSAZURE_STORAGE_PROXY, TESTS_ZEND_SERVICE_WINDOWSAZURE_STORAGE_PROXY_PORT, TESTS_ZEND_SERVICE_WINDOWSAZURE_STORAGE_PROXY_CREDENTIALS);
         }
 
         return $storageClient;
     }
-
+    
     protected function createSessionHandler($storageInstance, $tableName)
     {
         $sessionHandler = new Zend_Service_WindowsAzure_SessionHandler(
@@ -105,15 +97,13 @@ class Zend_Service_WindowsAzure_SessionHandlerTest extends PHPUnit_Framework_Tes
         );
         return $sessionHandler;
     }
-
-    protected static $uniqId = 0;
-
+   
     protected function generateName()
     {
         self::$uniqId++;
         return TESTS_ZEND_SERVICE_WINDOWSAZURE_SESSIONHANDLER_TABLENAME_PREFIX . self::$uniqId;
     }
-
+    
     /**
      * Test register
      */
@@ -123,10 +113,10 @@ class Zend_Service_WindowsAzure_SessionHandlerTest extends PHPUnit_Framework_Tes
         $tableName = $this->generateName();
         $sessionHandler = $this->createSessionHandler($storageClient, $tableName);
         $result = $sessionHandler->register();
-
+        
         $this->assertTrue($result);
     }
-
+    
     /**
      * Test open
      */
@@ -138,11 +128,11 @@ class Zend_Service_WindowsAzure_SessionHandlerTest extends PHPUnit_Framework_Tes
         $result = $sessionHandler->open();
 
         $this->assertTrue($result);
-
+        
         $verifyResult = $storageClient->listTables();
         $this->assertEquals($tableName, $verifyResult[0]->Name);
     }
-
+    
     /**
      * Test close
      */
@@ -153,10 +143,10 @@ class Zend_Service_WindowsAzure_SessionHandlerTest extends PHPUnit_Framework_Tes
         $sessionHandler = $this->createSessionHandler($storageClient, $tableName);
         $sessionHandler->open();
         $result = $sessionHandler->close();
-
+        
         $this->assertTrue($result);
     }
-
+    
     /**
      * Test read
      */
@@ -166,34 +156,37 @@ class Zend_Service_WindowsAzure_SessionHandlerTest extends PHPUnit_Framework_Tes
         $tableName = $this->generateName();
         $sessionHandler = $this->createSessionHandler($storageClient, $tableName);
         $sessionHandler->open();
-
+        
         $sessionId = $this->session_id();
         $sessionData = serialize( 'PHPAzure' );
         $sessionHandler->write($sessionId, $sessionData);
-
+        
         $result = unserialize( $sessionHandler->read($sessionId) );
-
+        
         $this->assertEquals('PHPAzure', $result);
     }
-
+    
     /**
      * Test write
      */
     public function testWrite()
     {
-        $storageClient = $this->createStorageInstance();
-        $tableName = $this->generateName();
-        $sessionHandler = $this->createSessionHandler($storageClient, $tableName);
-        $sessionHandler->open();
-
-        $sessionId = $this->session_id();
-        $sessionData = serialize( 'PHPAzure' );
-        $sessionHandler->write($sessionId, $sessionData);
-
-        $verifyResult = $storageClient->retrieveEntities($tableName);
-        $this->assertEquals(1, count($verifyResult));
+        if (TESTS_ZEND_SERVICE_WINDOWSAZURE_SESSIONHANDLER_RUNTESTS) {
+            $storageClient = $this->createStorageInstance();
+            $tableName = $this->generateName();
+            $sessionHandler = $this->createSessionHandler($storageClient, $tableName);
+            $sessionHandler->open();
+            
+            $sessionId = $this->session_id();
+            $sessionData = serialize( 'PHPAzure' );
+            $sessionHandler->write($sessionId, $sessionData);
+            
+            
+            $verifyResult = $storageClient->retrieveEntities($tableName);
+            $this->assertEquals(1, count($verifyResult));
+        }
     }
-
+    
     /**
      * Test destroy
      */
@@ -203,18 +196,18 @@ class Zend_Service_WindowsAzure_SessionHandlerTest extends PHPUnit_Framework_Tes
         $tableName = $this->generateName();
         $sessionHandler = $this->createSessionHandler($storageClient, $tableName);
         $sessionHandler->open();
-
+        
         $sessionId = $this->session_id();
         $sessionData = serialize( 'PHPAzure' );
         $sessionHandler->write($sessionId, $sessionData);
-
+        
         $result = $sessionHandler->destroy($sessionId);
         $this->assertTrue($result);
-
+        
         $verifyResult = $storageClient->retrieveEntities($tableName);
         $this->assertEquals(0, count($verifyResult));
     }
-
+    
     /**
      * Test gc
      */
@@ -224,16 +217,16 @@ class Zend_Service_WindowsAzure_SessionHandlerTest extends PHPUnit_Framework_Tes
         $tableName = $this->generateName();
         $sessionHandler = $this->createSessionHandler($storageClient, $tableName);
         $sessionHandler->open();
-
+        
         $sessionId = $this->session_id();
         $sessionData = serialize( 'PHPAzure' );
         $sessionHandler->write($sessionId, $sessionData);
-
+        
         sleep(1); // let time() tick
-
+        
         $result = $sessionHandler->gc(0);
         $this->assertTrue($result);
-
+        
         $verifyResult = $storageClient->retrieveEntities($tableName);
         $this->assertEquals(0, count($verifyResult));
     }
@@ -244,7 +237,3 @@ class Zend_Service_WindowsAzure_SessionHandlerTest extends PHPUnit_Framework_Tes
     }
 }
 
-// Call Zend_Service_WindowsAzure_SessionHandlerTest::main() if this source file is executed directly.
-if (PHPUnit_MAIN_METHOD == "Zend_Service_WindowsAzure_SessionHandlerTest::main") {
-    Zend_Service_WindowsAzure_SessionHandlerTest::main();
-}
